just use size_t ... environ.h is in Util/genutil, which has not yet been made

+++ Util/units/HeapTracked.C
@@ -14,7 +14,7 @@ using namespace std;
 // static ensures read-only access via get_heap_queue_size
 static vector<const void*> heap_addresses;
 
-uint64 Reference::HeapTracked::get_heap_queue_size ()
+size_t Reference::HeapTracked::get_heap_queue_size ()
 {
   return heap_addresses.size();
 }
